requirements.txt
Flask==3.1.2
Werkzeug==3.1.3
Pillow==12.0.0

